fix in postprocess

+++ roscausal_data/pp_scripts/postprocess.py
@@ -24,9 +24,8 @@ def wrap(angle, lower_bound, upper_bound):
 
     return wrapped_angle
 
-
 class Agent():
-    def __init__(self, name, x, y, theta, v, omega) -> None:
+    def __init__(self, name, x, y, theta = None, v = None, omega = None):
         self.name = name
         self.x = x
         self.y = y
@@ -48,7 +47,7 @@ class Agent():
     
     def dv(self, t):
         """
-        Decomposed velocity
+        decomposed velocity
 
         Args:
             t (int): time index
@@ -71,10 +70,9 @@ class Agent():
         """
         return self.p(t).distance(obs.p(t))
     
-    
     def heading(self, t, obj):
         """
-        Heading angle
+        heading angle
         
         Args:
             t (int): time index
@@ -87,28 +85,23 @@ class Agent():
         angle = wrap(2*np.pi - wrap(math.atan2(obj.p(t).y - self.p(t).y, obj.p(t).x - self.p(t).x) - wrap(self.theta[t], 0, 2*np.pi), 0, 2*np.pi), -np.pi, np.pi)
         return angle
     
-    
     def risk(self, t, obs):
         """
         Risk
 
         Args:
-            t (int): time index
-            obs (Agent): obstacle
-
-        Returns:
-            float: risk value
+            robot (RobotState): robot state
+            people (TrackedPersons): tracked people
         """
-
         
-        risk = self.v[t]
+        risk = self.v[t]/10 + np.random.normal(0, 0.02)
         collision = False
         
         # Calculate relative velocity vector
         Vrel = Point(obs.dv(t).x - self.dv(t).x, obs.dv(t).y - self.dv(t).y)
             
         # Compute the slope of the line AB and line PAB ⊥ AB
-        slope_AB = (obs.p(t).y - self.p(t).y) / (obs.p(t).x - self.p(t).x)  # Rise over run
+        slope_AB = (obs.p(t).y - self.p(t).y) / (obs.p(t).x - self.p(t).x) 
         slope_PAB = -1 / slope_AB
         # Choose distances to the left and right of point B
 
@@ -124,9 +117,9 @@ class Agent():
         cone = Polygon([cone_origin, left, right])
         
         P = Point(cone_origin.x + self.dv(t).x, cone_origin.y + self.dv(t).y)
-        collision = P.within(cone) and self.p(t).distance(obs.p(t)) < SAFE_DIST
+        collision = P.within(cone) and self.dist(t, obs) < SAFE_DIST
         if collision:
-            time_collision_measure = self.p(t).distance(obs.p(t)) / math.sqrt(Vrel.x**2 + Vrel.y**2)
+            time_collision_measure = self.dist(t, obs) / math.sqrt(Vrel.x**2 + Vrel.y**2)
             steering_effort_measure = min(P.distance(LineString([cone_origin, left])), P.distance(LineString([cone_origin, right])))           
             risk = risk + 1/time_collision_measure + steering_effort_measure
                     
@@ -142,6 +135,7 @@ if __name__ == '__main__':
     parser.add_argument("--obs_size", help="Obstacle increase size (radius) for risk calculation")
     parser.add_argument("--safe_dist", help="Safety distance for risk calculation")
     parser.add_argument("--sel_agent", help="Selected agent")
+    parser.add_argument("--delete_traj", help="If True. Delete traj file")
     args = parser.parse_args()
 
     CSV = args.csv
@@ -149,6 +143,8 @@ if __name__ == '__main__':
     PP_DATA_DIR = args.pp_data_dir
     OBS_SIZE = float(args.obs_size)
     SAFE_DIST = float(args.safe_dist)
+    SEL_ID = str(args.sel_agent)
+    DELETE_TRAJ = True if str(args.sel_agent) == 'True' or str(args.sel_agent) == 'true' else False
     
     INPUT_CSV = DATA_DIR + '/' + CSV
     OUTPUT_CSV = PP_DATA_DIR + '/' + CSV
@@ -157,13 +153,14 @@ if __name__ == '__main__':
     data = pd.read_csv(INPUT_CSV)
     
     R = Agent("R", data["r_x"], data["r_y"], data["r_{\theta}"], data["r_v"], data["r_{\omega}"])
-    H = Agent("H", data["h_x"], data["h_y"], data["h_{\theta}"], data["h_v"], data["h_{\omega}"])
-    RG = Agent("RG", data["r_{gx}"], data["r_{gy}"], np.zeros_like(data["r_{gy}"]), np.zeros_like(data["r_{gy}"]), np.zeros_like(data["r_{gy}"]))
-    HG = Agent("HG", data["h_{gx}"], data["h_{gy}"], np.zeros_like(data["r_{gy}"]), np.zeros_like(data["r_{gy}"]), np.zeros_like(data["r_{gy}"]))
+    H = Agent("H", data["h_" + SEL_ID + "_x"], data["h_" + SEL_ID + "_y"], data["h_" + SEL_ID + "_{\theta}"], data["h_" + SEL_ID + "_v"], data["h_" + SEL_ID + "_{\omega}"])
+    RG = Agent("RG", data["r_{gx}"], data["r_{gy}"])
+    HG = Agent("HG", data["h_" + SEL_ID + "_{gx}"], data["h_" + SEL_ID + "_{gy}"])
             
     df = pd.DataFrame(columns=["h_v", r"h_{\theta}", r"h_{\theta_{g}}", "h_{d_g}", "h_{risk}", r"h_{\omega}", r"h_{d_{obs}}"])  
     
-    for i in range(1, len(data)):
+    I0 = 1
+    for i in range(I0, len(data)):
                 
         df.loc[i] = {"h_v" : H.v[i],
                      r"h_{\theta}" : H.theta[i], 
@@ -175,8 +172,10 @@ if __name__ == '__main__':
                      }
 
     # Save the processed data to another CSV file
-    df = df[1:]
+    df = df[I0:-1]
+
     df.to_csv(OUTPUT_CSV, index=False)
     
-    os.chmod(INPUT_CSV, stat.S_IWUSR | stat.S_IRUSR | stat.S_IXUSR)
-    os.remove(INPUT_CSV)
+    if DELETE_TRAJ:
+        os.chmod(INPUT_CSV, stat.S_IWUSR | stat.S_IRUSR | stat.S_IXUSR)
+        os.remove(INPUT_CSV)
